move legacy types to llegacy directory
