+++ src/example-types/legacy/rangeStats.js
@@ -1,5 +1,5 @@
 let _ = require('lodash/fp')
-let esTwoLevel = require('./esTwoLevelAggregation').result
+let esTwoLevel = require('../esTwoLevelAggregation').result
 
 module.exports = {
   validContext: node => !!(node.key_field && node.value_field && node.ranges),
